moved somethings around to make it more logical
