+++ package.json
@@ -2,15 +2,14 @@
     "name": "vigilant-fiesta",
     "version": "0.1.0",
     "description": "A hobby project creating a simple ecosystem sim.",
-    "main": "dist/Main.js",
+    "main": "dist/main.js",
     "scripts": {
         "postinstall": "electron-builder install-app-deps",
         "pack": "cross-env DEBUG=electron-builder electron-builder build",
         "prestart": "npm install && npm run build && npm run lint",
         "start": "cross-env NODE_ENV=development npm run electron",
-        "cleanup": "rimraf dist/*.js && rimraf dis/*.html",
-        "copy-html": "mkdirp dist/ && cpx src/renderer/index.html dist/",
-        "build": "npm run cleanup && rollup -c && npm run copy-html",
+        "cleanup": "rimraf dist/*.js",
+        "build": "npm run cleanup && rollup -c",
         "tsc": "tsc",
         "unit-tests": "cross-env NODE_ENV=test electron-mocha --require ts-node/register test/**/*_spec.ts",
         "test": "npm run lint && npm run unit-tests",
@@ -67,8 +66,8 @@
         "files": [
             "dist/",
             "dist/index.html",
-            "dist/Main.js",
-            "dist/Renderer.js"
+            "dist/main.js",
+            "dist/index.js"
         ]
     }
 }
