+++ rollup.config.js
@@ -3,9 +3,9 @@ import { terser } from "rollup-plugin-terser";
 
 export default [
     {
-        input: 'src/main/Main.ts',
+        input: 'src/coreElectron/Main.ts',
         output: {
-            file: 'dist/Main.js',
+            file: 'dist/main.js',
             format: 'cjs'
         },
         plugins: [
@@ -18,9 +18,9 @@ export default [
         external: ['electron']
     },
     {
-        input: 'src/renderer/Renderer.tsx',
+        input: 'src/pages/Index.tsx',
         output: {
-            file: 'dist/Renderer.js',
+            file: 'dist/index.js',
             format: 'cjs'
         },
         plugins: [
