--- src/renderer/index.html
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Document</title>
-    <style>
-        div#react-entry-point, html, body {
-            margin: 0;
-            height: 100%;
-            overflow: hidden;
-        }
-        html * {
-            background: var(--canvas-background-color);
-        }
-        .application, body canvas#mainboard {
-            margin: auto;
-            width: 100%;
-            height: 100%;
-        }
-        .ui {
-            border-radius: 5px;
-        }
-        .speedControls {
-            background-color: rgba(255, 255, 255, 0.233);
-            padding: 5px;
-            margin: 10px;
-            bottom: 0px;
-            right: 0px;
-            position: absolute;
-        }
-        .speedControls .smallButton {
-            width: 50px;
-        }
-        .smallButton {
-            background-color: rgb(255, 0, 0);
-            border: 0px;
-            margin: 5px;
-        }
-    </style>
-    <script src="./Renderer.js"></script>
-</head>
-<body>
-    <div id="react-entry-point"></div>
-</body>
-</html>
