+++ tslint.json
@@ -21,7 +21,7 @@
         "label-position": true,
         "max-line-length": [
             true,
-            200
+            100
         ],
         "member-access": false,
         "member-ordering": [
